TEC-2856 Add defaultProps. Tweak css
