+++ example/src/mocks/mock.js
@@ -532,51 +532,6 @@ export const mockLineItems = [{
   'variantId': 3108,
   'variantStock': 87 }]
 
-export const desktopLinks = {
-  'support': {
-    'faq': { url: '/faqs' },
-    'shippingAndReturns': { url: '/faqs' },
-    'materials': { url: '/materials' },
-    'stylingHelp': { url: '/info/styling-appointments' },
-    'ringSizer': { url: '/ring-sizer' },
-    'contactUs': { url: '/contact' }
-  },
-  'company': {
-    'aboutUs': { url: '/about-us' },
-    'careers': { url: 'https://jobs.lever.co/mejuri', isExternal: true }
-  },
-  'visitUs': {
-    'losAngeles': { url: '/irl-coming-soon' },
-    'newYork': { url: '/irl/new-york' },
-    'toronto': { url: '/irl/toronto' },
-    'piercingStudio': { url: '/info/piercing-studio' }
-  }
-}
-
-export const mobileLinks = {
-  'support': {
-    'faq': { url: '/faqs' },
-    'shippingAndReturns': { url: '/faqs' },
-    'materials': { url: '/materials' },
-    'stylingHelp': { url: '/info/styling-appointments' },
-    'ringSizer': { url: '/ring-sizer' },
-    'careers': { url: 'https://jobs.lever.co/mejuri', isExternal: true }
-  },
-  'visitUs': {
-    'losAngeles': { url: '/irl-coming-soon' },
-    'newYork': { url: '/irl/new-york' },
-    'toronto': { url: '/irl/toronto' },
-    'piercingStudio': { url: '/info/piercing-studio' }
-  }
-}
-
-export const socialLinks = [
-  { tag: 'Facebook', link: 'https://facebook.com/mejuri' },
-  { tag: 'Instagram', link: 'https://instagram.com/mejuri' },
-  { tag: 'Pinterest', link: 'https://pinterest.com/mejuri' },
-  { tag: 'Twitter', link: 'https://twitter.com/intent/user?screen_name=mejuri' }
-]
-
 export const mockSummary = { 'lineItems':
 [{ 'id': 507708,
   'quantity': 2,
@@ -644,7 +599,3 @@ export const mockSummary = { 'lineItems':
   'shipping': null,
   'adjustments': [{ 'id': 591401, 'label': 'Promotion (patotest)', 'displayAmount': '-$23.6', 'amount': '-23.6', 'eligible': true }],
   'total': '$94.4' } }
-
-
-
-
