+++ package.json
@@ -33,6 +33,7 @@
     "humps": "^2.0.1",
     "lodash.debounce": "^4.0.8",
     "lodash.get": "^4.4.2",
+    "lodash.map": "^4.6.0",
     "lodash.throttle": "^4.1.1",
     "nookies": "^2.3.0"
   },
