+++ src/components/Footer/defaultProps.js
@@ -1,0 +1,44 @@
+export const desktopLinks = {
+  support: {
+    faq: { url: '/faqs' },
+    shippingAndReturns: { url: '/faqs' },
+    materials: { url: '/materials' },
+    stylingHelp: { url: '/info/styling-appointments' },
+    ringSizer: { url: '/ring-sizer' },
+    contactUs: { url: '/contact' }
+  },
+  company: {
+    aboutUs: { url: '/about-us' },
+    careers: { url: 'https://jobs.lever.co/mejuri', isExternal: true }
+  },
+  visitUs: {
+    losAngeles: { url: '/irl-coming-soon' },
+    newYork: { url: '/irl/new-york' },
+    toronto: { url: '/irl/toronto' },
+    piercingStudio: { url: '/info/piercing-studio' }
+  }
+}
+
+export const mobileLinks = {
+  support: {
+    faq: { url: '/faqs' },
+    shippingAndReturns: { url: '/faqs' },
+    materials: { url: '/materials' },
+    stylingHelp: { url: '/info/styling-appointments' },
+    ringSizer: { url: '/ring-sizer' },
+    careers: { url: 'https://jobs.lever.co/mejuri', isExternal: true }
+  },
+  visitUs: {
+    losAngeles: { url: '/irl-coming-soon' },
+    newYork: { url: '/irl/new-york' },
+    toronto: { url: '/irl/toronto' },
+    piercingStudio: { url: '/info/piercing-studio' }
+  }
+}
+
+export const socialLinks = [
+  { tag: 'Facebook', link: 'https://facebook.com/mejuri' },
+  { tag: 'Instagram', link: 'https://instagram.com/mejuri' },
+  { tag: 'Pinterest', link: 'https://pinterest.com/mejuri' },
+  { tag: 'Twitter', link: 'https://twitter.com/intent/user?screen_name=mejuri' }
+]
